+++ src/exact_0804cc28.c
@@ -6,7 +6,7 @@
 #include "sio.h"
 
 /* prototypes for same-file helpers called by this run */
-void Sio_StartBoxTransitionClose(u16 a, ProcPtr parent);
+void Sio_StartBoxTransitionClose(int a, ProcPtr parent);
 
 //! FE8U = 0x0804BE88
 void Sio_StartBoxTransitionCloseDefault(ProcPtr parent)
